+++ skrub/_reporting/_table_report.py
@@ -3,6 +3,7 @@ import functools
 import json
 from pathlib import Path
 
+from .. import _dataframe as sbd
 from ._html import to_html
 from ._serve import open_in_browser
 from ._summarize import summarize_dataframe
@@ -42,6 +43,10 @@ class TableReport:
 
         * verbose = 1 prints how many columns have been processed so far.
         * verbose = 0 silences the output.
+    max_plot_columns : int, default=30
+        Maximum number of columns for which plots should be generated.
+        If the number of columns in the dataframe is greater than this value,
+        the plots will not be generated. If None, all columns will be plotted.
 
     See Also
     --------
@@ -115,6 +120,7 @@ class TableReport:
         title=None,
         column_filters=None,
         verbose=1,
+        max_plot_columns=30,
     ):
         n_rows = max(1, n_rows)
         self._summary_kwargs = {
@@ -127,6 +133,7 @@ class TableReport:
         self.column_filters = column_filters
         self.dataframe = dataframe
         self.verbose = verbose
+        self.max_plot_columns = max_plot_columns
 
     def __repr__(self):
         return f"<{self.__class__.__name__}: use .open() to display>"
@@ -143,11 +150,15 @@ class TableReport:
             self.dataframe, with_plots=False, title=self.title, **self._summary_kwargs
         )
 
-    @property
-    def _any_summary(self):
-        if "_summary_with_plots" in self.__dict__:
-            return self._summary_with_plots
-        return self._summary_without_plots
+    def _get_summary(self):
+        if self.max_plot_columns is None:
+            summary = self._summary_with_plots
+        elif self.max_plot_columns >= sbd.shape(self.dataframe)[1]:
+            summary = self._summary_with_plots
+        else:
+            summary = self._summary_without_plots
+
+        return summary
 
     def html(self):
         """Get the report as a full HTML page.
@@ -158,7 +169,7 @@ class TableReport:
             The HTML page.
         """
         return to_html(
-            self._summary_with_plots,
+            self._get_summary(),
             standalone=True,
             column_filters=self.column_filters,
         )
@@ -172,7 +183,7 @@ class TableReport:
             The HTML snippet.
         """
         return to_html(
-            self._summary_with_plots,
+            self._get_summary(),
             standalone=False,
             column_filters=self.column_filters,
         )