+++ CHANGES.rst
@@ -62,6 +62,12 @@ Changes
 - Optimize the :class:`StringEncoder`: lower memory footprint and faster execution in some cases.
   :pr:`1248` by :user:`Gaël Varoquaux <gaelvaroquaux>`
 
+
+- A new parameter ``max_plot_columns`` has been added to the :class:`TableReport` and :func:`patch_display`
+  to skip column plots when the number of columns exceeds the specified value.
+  :pr:`1255` by :user:`Priscilla Baah<priscilla-b>`.
+
+=======
 Bug fixes
 ---------
 - :class:`StringEncoder` now works correctly in presence of null values.
@@ -70,6 +76,7 @@ Bug fixes
 - The :meth:`TableVectorizer.get_feature_names_out` method now works when used in a
   scikit-learn pipeline by exposing the `input_features` parameter.
   :pr:`1258` by :user:`Guillaume Lemaitre <glemaitre>`.
+
 
 Release 0.5.1
 =============